requirements.txt
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
easyocr==1.7.2
filelock==3.16.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
imageio==2.35.1
Jinja2==3.1.4
jiter==0.5.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lazy_loader==0.4
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.8.3
networkx==3.3
ninja==1.11.1.1
numba==0.60.0
numpy==2.0.2
openai==1.47.1
opencv-contrib-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.3
pdf2image==1.17.0
pdf2table==0.1.3
pillow==10.4.0
polars==1.8.2
poppler-utils==0.1.0
protobuf==5.28.2
psycopg2-binary==2.9.9
pyarrow==17.0.0
pyclipper==1.3.0.post5
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pypdf==5.0.0
PyPDF2==3.0.1
pytesseract==0.3.13
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.1
setuptools==75.1.0
shapely==2.0.6
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
streamlit==1.38.0
streamlit-drawable-canvas==0.9.3
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
tifffile==2024.9.20
tokenizers==0.19.1
toml==0.10.2
torch==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==4.0.2
